golangci: Ignore gci errors in CGO files.
It is possible to change these files to keep them in line with gci but
that requires moving all C code to the beginning of the import block,
then importing C, and then importing everything else. I think this looks
less clean than how we currently do it, so ignoring these files.

Signed-off-by: Mark Laing <mark.laing@canonical.com>