lxd: Check view permissions in `GET /1.0/storage-pools`.
`can_view` on a `storage_pool` is defined as `can_view` from `server`.
`can_view` on `server` is a type-bound public access. This means that
no extra restriction is being applied here, all authenticated identities
will still be able to view all storage pools. The difference is in
OpenFGA validation when it receives a check request. It will validate
that only `identity` (or `service_account`) types can have this
relation.

Signed-off-by: Mark Laing <mark.laing@canonical.com>