lxd: Move CA check into `certificateValidate` method.
Whether a new client is being trusted via secret, or if an
administrator is adding or updating the certificate directly,
the certificate must be signed by the CA if set.

Note that this is *not* a security issue, because we validate
that client certificates have been signed by the CA when authenticating.
This just meant that it was possible for an admin to create a certificate
that would be invalid.

Signed-off-by: Mark Laing <mark.laing@canonical.com>